requirement.txt
streamlit
torch
torch-geometric
rdkit-pypi
matplotlib
networkx
pandas
scikit-learn
